CMS-12324 Hide busy indicator initial doc load
Call hide busy indicator when the src of the CMS iframe reloads.

When wicket start a xhr request then it fires a
Wicket.Event.Topic.AJAX_CALL_BEFORE_SEND event. Ajaxindicator.js adds a
listener to that event to make a call to the navapp to show the
busyindicator.
When wicket completes a xhr request it sends a
Wicket.EVENT.Topic.AJAX_CALL_COMPLETE event. Ajaxindicator adds a
listener to that even to make a call to the navapp to hide the
busyindicator. If the iframe reloads in between those two events, the
network tab in the browser still shows a 200, but any handlers on that
xhr are removed. As a result the AJAX_CALL_COMPLETE event is not fired.

As a workaround before call hideBusyIndicator before reloading the page.
This assumes that one wicket xhr call has started when the iframe
is reloaded.