REPO-2287 /repository only allow userrole xm.repository-browser.user
Not everybody should be allowed to login into the repository servlet
with a valid jcr session. The repository servlet should only be
accessible to those who have a role that says they are allowed to do so

for this, introduce the xm.repository-browser.user role. By default,
admin user and admin group gets this role

The role does not give any jcr node privileges, but is only used for
checking whether the user is in a role to access the repository servlet