CHANNELMGR-2440 Clean channel upon navigate
When comparing the sequence of methods invocations made when the channel
was closed with the menu inside the experience manager iframe with
selecting the "root" item of the breadcrumb a call to "clear channel"
was missing when navigating to the overview via the breadcrumb.

Looking at the commit for CHANNELMGR-2398 it's also clear that the
original breadcrumb also cleared the channel before navigating to the
overview.

There was no way yet of communicating TO the channel manager iframe.
However the channel manager already registred itself: See IframePanel.js
inside the cms code.

Navigate was not called, that is fix in CMS-12360.

Now that the iframe receives a navigate call it can clear the channel,
however, it should not update the nav location, because that interferes
with call logic already in place for updating the breadcrumb. That
should not be touched.