HIPPLUG-1602 Add required compound validator
When a field of type hippo:compound is marked as "required", is should
at least contain a single instance of that field. For built-in compound
fields, the Wicket editor takes cares of this, as it ensures there is at
least a single compound field in the document.

For content-blocks however, there can be zero instances. In case it is
marked as "required", this should render a violation in context of the
content-block.