REPO-2285 improve autoexport of security configuration which also can (be expected to) be modified in production
Autoexport by default doesn't 'work' for category system nodes and properties as those intentionally are supposed
to be (only) modified in production.
However, security *configuration* (not users, groups) typically this is setup/defined (first) by developers, while
still allowing to be modified and 'augmented' in production, for example using the CMS security admin UI.

With the current practice, such runtime modifications *should* be merged back into project bootstrap configuration,
manually or by using the Configuration Verifier, until there is more integrated tooling to automate and synchronize
such runtime (security configuration) changes.

To better support the main use-case, developer managed and configured security configuration, specific security
configuration properties are *categorized* as system but *also* will be autoexported using a (hardcoded) override
for the following properties of a hipposys:authrole node:
  - hipposys:userrole
  - hipposys:users
  - hipposys:groups
and for the hipposys:userroles property of the following node types:
  - hipposys:user
  - hipposys:externaluser
  - hipposys:group
  - hipposys:externalgroup
All product provided configuration for the above properties will be categorized system, and for multi-valued properties
with .meta:add-new-system-values (see HCM-283), and autoexport now will also export this .meta configuration.

In addition, for role and userrole the residual childnode category is set/restored to config instead of system to also
primarly support the developer use-case, instead of the currently less prominent runtime changes use-case.
For the latter, again and like for runtime changes of the above mentioned properties and authroles, a developer/admin
*should* use the Configuration Verifier to merge back runtime production changes *before* deploying (then) destructive
changes made through the project bootstrap configuration!