REPO-2284 in case a FacetFilter cannot be created, throw IllegalArgExc
Only IllegalArgumentExceptions are handled gracefully, however, there
can be other exception like for example ArrayOutOfBounds in case the
facet filter is for example:

myproject:color!=

In this case, the fac nav node cannot be loaded and also not fixed any
more. Instead, we should catch any exception and wrap it in a
IllegalArgumentException which is handled gracefully.