HSTTWO-4686 [forward port 14.1] Correct the workflow handling for committing a draft
After creating a new document with workflow, it is a draft document. To
create a preview for this one, we *should not* set properties on the
document with the user session but have this handled by the workflow:
First of all, the workflow knows how to commit the editable instance and
secondly, in general, users only need readwrite privilege to draft
documents meaning they *cannot* change properties like the hippostd:state
into a different state since they are not allowed to write to nodes
with that new state!

(cherry picked from commit 232953492a3931927a9c358e8ac8869d4f54b633)