CMS-12383 Add relative favicon url to AppSettings
Add relative favicon url to app settings, so that the navapp can prepend
the "base url", so that the favicon can be fetched from the cms.
Currently the favicon is fetched from the angular "base href" and it
cannot be found.

TLDR;
When a url is relative the resource is fetched from the base href, which
is the location from where the navapp is served. We cannot use an
absolute url, because the tomcat instance is not aware of any reverse
proxy setup by the web server.
The navigation application however is aware of its current url and can
prepend the "base url".
The faviconUrl is for example
"./wicket/resource/org.hippoecm.frontend.service.WicketFaviconServiceImpl/cms-icon.png?ver=1574774708000".
Now if a client accesses cms.example.com, the favicon can be found at
cms.example.com/wicket/resource/org.hippoecm.frontend.service.WicketFaviconServiceImpl/cms-icon.png?ver=1574774708000.
The reverse proxy config will map to a tomcat node, e.g.
ip:port/cms/wicket/resource/org.hippoecm.frontend.service.WicketFaviconServiceImpl/cms-icon.png?ver=1574774708000
If we would provide the absolute url of the tomcat, the client won't be
able to find the resource, because the faviconUrl is passed on to the
navigation application and it reads it client side.

Add cms-icon.png and console-icon.png to the resources in the same
package as the WicketFaviconServiceImpl, because they were not present
on the previous location, the package of the Main class.
Include resource png in the maven build, so they can be found.

Check if the file exists instead of if the resource is null, because the
resource can be non-null, but still point to a url that cannot be found.

Introduce WicketFaviconServiceImpl, because the business logic was in
Main, LoginPlugin and PluginPage.