HSTTWO-4688 Correct sitemap item level authorization combination with mount
A visitor now needs to pass

- the sitemap item authorization
- the mount authorization

So for example when the mount is authorized to only be accessible for
users in role 'staff', then if a sitemap item has the requirement
'superstaff' role, then the user must be in role 'staff' and in role
'superstaff'.

A sitemap item can override the authenticated to set it to false. Then
that sitemap item is not authenticated, even if the rest of the mount
is authenticated.

If the sitempap item has authenticated = true, but it does not have
any roles/users configured, there are two use cases:

1) If the mount does not have authenticated = true, it means the sitemap
item is not accessible for anyone
2) If the mount does have authenticated = true, we check the roles/users
of the mount