CMS-12401 Show error on non-response wicket xhr
Show error page in navigation application when interacting with the cms
and the server does not respond.

Send an onError message [penpall] with a 500 error code and a message to the
navigation application when an xhr fails.

Add a custom javascript header item, adds the no-response.js

Add Hippo.onError global method so it can be called from the javascript
that is included in the header item.

In cms versions before version 14 we did not show any error page if a
xhr call failed. Now that the navigation application loads the cms in an
iframe, the navigation application updates the breadcrumb and the menu
item, but the cms does not respond anymore. This is inconsistent from a
ux point of view.

Wicket uses jquery to send xhr calls and has a number of callbacks, see
wicket-core-7.14.0 /org/apache/wicket/ajax/res/js/wicket-ajax-jquery.js:1119.

When the wicket xhr call fails it sends a AJAX_CALL_FAILURE event.