REPO-2297 jcr:read privilege is returned when user can read absPath
When invoking

public DomainInfoPrivilege[] getPrivileges(String absPath)

privilege jcr:read was not returned in case of implicit read access on
absPath (and no explicit domain grants jcr:read for the user on the path).

This is incorrect since although the read is implicit, the user has the
privilege to read the node.

Therefor the code now explicitly checks whether jcr:read is present and
if not, check canRead which in case it returns true, add the jcr:read
privilege