CHANNELMGR-2445 Do not connect to child when setting location
The IFrame panel is an ExtJs class that can be used to create a
panel containing an IFrame. Not all users of (instances of) this
class require a connection to a child inside the iframe.
It's up to a subclass of IFramePanel or code using the panel to
decide if/when connecting to a child in the iframe makes sense.