CMS-12419 Open document with behaviour
Open document with behaviour instead of by appending parameters to the
url.

Wait with opening a document until the document perspective is loaded.
The javascript function Hippo.openDocumentById is added as last header
item, so by then the server side part of the document editor for sure is
ready to open the editor.

Open a document by handle identifier or by handle path depending on the
supplied appPath.