HSTTWO-4690 Lazily generate bean for hippo:compound
No dynamic bean was generated for a compound type, unless this type was
registered in the nodetypes section of some document type. With this fix
we allow creation of the compound dynamic bean in a lazy way (when requested
via objectconverter's getObject method).
A good example when this scenario happens is when using content blocks

+ tiny improvements