CMS-12351 Only reload top window if url has iframe search param
This is a work-around, but it is necessary to make sure that the
logout message (yellow info bar in login panel) is not lost due
to the reload of the top window.

We tried to implement logout from Autho0 by means of an XHR call
triggered from javascript code inside of the CMS iframe but that
didn't work (for unclear reasons). As a consequence we have to do
a logout by letting the server initiate the logout process so that
the browser is redirected several times between Auth0 and the CMS
to complete the full logout. However, it is the navapp that requested
the CMS iframe to logout. In response to this request the CMS returns
a resolved promise. We can't to wait with resolving due to the redirects.
So the navapp has to wait a bit (1 second) to be (reasonably) sure that
the logout in the CMS iframe has completed. After that the navapp
redirects the top window to the login page of the CMS with a query parameter
loginmessage. This loginmessage key is used to lookup the message to display
the yellow info bar in the login form.

So then why is this of javascript even needed you might ask. Because of the
ConcurrentLoginFilter. This filter invalidates a server-side session if a
login occurs with the same credentials as the session (does not work cluster
wide though so it's a rather buggy 'feature'). You probably have experienced
this behaviour yourself too on shared test environments. If a session is
concurrently logged out then the login screen is rendered in the CMS iframe.
So then the user could login again which would result in a navapp containing
a navapp containing the cms. The injected javascript will prevent this by
immediately reloading the top window. This works because the iframe URL does
not change so it still has the iframe query parameter. If the login is
initiated by the navapp then the iframe URL does change due to the
redirects and then the iframe query parameter is no longer present. The
user will briefly (at most 1 second) see the login page inside of the
iframe, but then the browser will refresh (as described above).