CMS-12441 Make backwards compatible, configurable
Only build an editor if the node to be edited is not one of the
blacklisted nodetypes.  When an editor is created for a nodetype, the
editor is first created.  Upon initialization of the editor is registers
itself as a service. That registration is picked up by the
EditorTabsPlugin ( it has a service tracker ) and a document tab is
added. If an editor is created for a node that does not have any default
document type and no namespace config for the editor templates, the
editor panel remains empty. If the workflow cannot add any menu items as
well it appears as if only a tab is added.  A full refactoring is out
scope. That would include registering an editor only if a valid template
exists and let custom plugin configure their own IEditorFactory
instance.  So let's settle for the solution where a blacklist of node
type names can be configured. Might a custom plugin have a editor for a
black listed type, it only has to delete the black listed node type name
from the list. E.g. if a custom editor would exist for a hippostd:folder
to manage certain properties on the folder node, that is still possible.

Remove the creation of an IEditorFactory in the EditorManagerPlugin by
instantiating the editor in a method, so that the contract of the
IEditorFactory is not broken. The contract is that the
IEditorFactory#newEditor return null if no editor can be created, while
the old instance in the EditorManagerPlugin threw an Exception is an
editor could not be created.

Use null checks instead of Optional to make sure the code is backwards
compatible, re-use node instance instead of creating a new node from the
session, format javadoc correctly and add descriptions. ( Thank for
spotting Michiel Eggermont ).

Migrate some java 5, 6 and 7 constructions to java 8.