HSTTWO-4693 Make sure Links are created differently for PageModelApi req
For a PageModelApi request the toUrlForm of HstLink objects must have
different behavior than for normal HST rendered webpages. The big
difference is that for normal HST rendered pages, all URLs in the
channel mgr should be relative to the current CMS host as the Channels
are rendered over the HOST of the CMS.

However, PageModelApi responses are consumed by an SPA that is rendered
server or client side. The URLs in the response of the PageModelApi
should not be absolute URLs (starting with /), since the SPA in the
browser typically runs on a different HOST than the CMS. Hence those
URLs should still be
1) Fully Qualified URLs including scheme for 'resource' and 'external'
links
2) Absolute links (starting '/') for SPA 'internal' links

To be able to better differentiate between requests that are for
PageModelApi requests, include this knowledge as a primary citizen of the
HstRequestContext