HSTTWO-4703 Support preflight (OPTIONS) request for all HST pipelines
Implemented behavior:

# For PMA (page model api) response, by default we add
Access-Control-Allow-Credentials: true
# For hst:responseheaders containing Access-Control-Allow-Origin: *,
we replace '*' in the response with the explicit request origin
(required by browsers in case of Access-Control-Allow-Credentials: true)
# If the HST does not contain Access-Control-Allow-Origin header but it
does contain hst:allowedorigins multi-valued property, we return
Access-Control-Allow-Origin: request-domain-host if the origin is
whitelisted. If the origin is not whitelisted, a 204 is returned
without other CORS headers
# If hst:responseheaders contains Access-Control-Allow-Origin  with as
value an explicit host, for example

Access-Control-Allow-Origin: http://www.example.org

we do not check hst:allowedorigins at all and just keep
http://www.example.org as Access-Control-Allow-Origin. If the current
request comes from a different domain, no further CORS headers are set,
since the domain is not allowed.
# The HST has a new service through which downstream modules/projects
can tell which headers (Access-Control-Allow-Headers) are allowed : This
is needed when for example an SPA does a request with a custom header :
in that case the browser first does an OPTIONS request to 'ask' the
server whether the request is allowed with the custom header
# The HST will for all NON-CXF (jaxrs) pipelines by default support
"HEAD", "GET", "POST" as allowed methods. These values are returned
explicitly in the Access-Control-Allow-Methods header (although not
needed perse since "HEAD", "GET", "POST" methods are always allowed)
# The HST will include by default ACCESS_CONTROL_MAX_AGE for the
preflight OPTIONS requests. We'll set it to 1 day by default for now
# The HST will include a header Vary: Origin which is very important to
avoid cached responses for preflight requests from different origins

Implementation details

- CorsSupportValve

All non-cxf related pipelines have a CorsSupportValve instance added as
the first processing valve. The CorsSupportValve handles OPTIONS
requests for non-cxf pipelines.

-hst:allowedorigins

On a node of type hst:virtualhost, a new property is present: 'hst:allowedorigins'. It is accessible in code via

org.hippoecm.hst.configuration.hosting.VirtualHost#getAllowedOrigins

- AccessControlAllowHeadersService

The HST platform webapp registers a new service, the
AccessControlAllowHeadersService. This service enables any downstream
project to contribute 'allowed headers' which are required for CORS
related requests. For example the targeting module informs via this
service that the customer header '_visitor' is allowed.