HSTTWO-4705 Classloader of ChannelInfo is shared loader opposed to the mixin
The mixin channel info classes must be loaded with the website class
loader. If there is no explicit/custom channel info class specified on
the hst:channel, the HST falls back to

org.hippoecm.hst.configuration.channel.ChannelInfo

which is from the shared lib. The classloader from the shared lib can
however not read the explicit mixins which are part of the website
classloader