HSTTWO-4708 Clean up dead code 'previewmode'
in the past, when previewing a channel outside the CMS, we'd first
switch to 'previewmode' opposed to the 'composer mode' : The only
difference would be that in 'composer mode'  the required html comments
for the channel mgr interactions were added.

However since CHANNELMGR-2199 the HST 'resty' endpoint for switching to
'preview mode' is never invoked any more, and preview is just rendered
including the html comments (which really doesn't harm).

Hence, let's remove the 'composer mode' cruft altogether since there is
at this moment no usage for it (also note that even in preview mode we'd
already added the page meta data comments any way)