HSTTWO-4709 Support non-local Channel Mgr Preview URL for a Channel
For SPA's rendered in the Channel Manager preview, it should be possible
that they are rendered in the channel mgr iframe from a different
hostname! Thus for example the cms runs on cms.example.com and the SPA
preview channel then needs to be possible to be rendered from
my.spa.example.com

As a result, when opening the channel in the channel mgr, we should be
able to specify *which* URL the channel mgr should open

A channel the configures hst:channelinfomixins to contain

org.hippoecm.hst.configuration.channel.PreviewURLChannelInfo

has now the possibility to store the property

org.hippoecm.hst.configuration.channel.PreviewURLChannelInfo_url

on the hst:channelinfo node. If such a property is available,
then the Channel Mngr should load that URL to show the preview
of the channel.

Note for now we have

@Parameter(name = PREVIEW_URL_PROPERTY_NAME, displayName = "Preview URL",
          hideInChannelManager = true)
String getPreviewURL();

The 'hideInChannelManager' means that the property is (not yet)
configurable via the channel settings. Note that the
PreviewURLChannelInfo however has to be present, otherwise the
property ....PreviewURLChannelInfo_url won't be kept after
changing channel settings