HSTTWO-4710 Bind JWT token to CmsSessionContext
For multiple reasons we need SPA requests for the page model api to go
to the same cluster node all the time, and that cluster node should also
have the http session of the logged in CMS user. The reasons for this are

1) Latency: When shuffling two component items around on the Channel
mgr preview, the next SPA page model api request should directly get
the updated result. the shuffling of two components is a request over
the host of the CMS like normal website channel mgr interactions are
handled. If the page model api request after the reshuffling hits
another cluster node, this cluster node might not yet have the changes,
resulting in the channel mgr preview not to reflect the changes. To
avoid this, we have to make sure the page model API requests hit the
same cluster node as the logged in CMS user is on.
2) In the Channel mgr, you can change the branch (project) to render,
or can can change the variant (alter ego) of the preview to be rendered:
Setting branch or variant is stored on the CmsSessionContext. Therefor,
via the Token the HST needs to be able to look up the CmsSessionContext.
The alternative would be that things which are stored in the cms session
context would have to get stored in the Token, and on every change, a
new Token would have to be generated. That is quite unhandy and complex
hence we better can support looking up the CmsSessionContext for a Token

Since the requirement above, we now bind the Token to the
CmsSessionContext : we set the subject of the token equal to the id of
the CmsSessionContext and introduced a TokenCmsSessionContextRegistry:
Via this registry, and existing CmsSessionContext can be retrieved
without requiring an http session, but just the CmsSessionContext id is
enough.

To make sure the TokenCmsSessionContextRegistry gets cleaned up once a
user logs out of the CMS, the use http session object listeners to
remove a token subject from the registry in case a user logs out or its
http session expires.