HSTTWO-4597 Server channel mgr preview over live host support
In case the request contains an "Authorization" header starting with
"Bearer", the HstDelegateeFilterBean will try to get an AccessToken
for the authorization: Assumed is a signed JWT token, where the same
cluster node running the HST has previously created and signed the
JWT token in JwtTokenServiceImpl#createToken

If the JWT token is valid (aka it is a valid signed Token and the
subject is tight to a valid CmsSessionContext via the
TokenCmsSessionContextRegistry, then the HST will render a
channel manager preview response (aka preview hst config, content
and including html comments). The response can be a normal HST
rendered website (html) or a Page Model API response (json)

As a result of the above, all the places where we use the
CmsSessionContext, we now need to try to get hold of the
CmsSessionContext via the AccessToken *or* via the HttpSession
for the rendering without AccessToken but the old school
website rendering in the channel mgr itself.

If the Channel Mgr opens a channel which should open an iframe
loading a (SPA) site from a different host because the channel
config contains

"org.hippoecm.hst.configuration.channel.PreviewURLChannelInfo_url"

then the HST will do the following redirects:

1) Channel is clicked in channel mgr overview, doing first a
_cmssessioncontext call
2) If the http session on the HST site webapp has not yet been
synced with the platform webapp CmsSessionContext, a redirect is
done to the cms webapp. After this, the cms webapp redirects
again to the site webapp which then establishes the CmsSessionContext
syncing. Note these requests are ALL over the host of the CMS!
3) After two has been done, for normal HST rendered websites
the preview of the channel is rendered. However IF there is a
specific 'PreviewURLChannelInfo_url' set, the HST will redirect
again to the specified URL including a query param for a
signed JWT (jws) token and a query param for 'serverid'
(possibly null if not serverid present)

After (3), typically an SPA is rendered invoking the HST
Page Model API with the token as header and the serverid
as header or query param (just what is most handy on proxy
level like nginx to support cluster node affinity)