HSTTWO-4703 Improvements after review feedback
After some more thoughts, most of the HST pipelines should not support
the OPTIONS request at all: Just like many websites (google, twitter,
reddit, github, etc), if there is no purpose for the OPTIONS request, it
is best to disallow the options request.

It is better to whitelist all the requests that should support CORS
requests (and thus support OPTIONS request) instead of by default
enable the OPTIONS request for URLs which do not support CORS requests
any way (other than GET, HEAD and POST).

Therefor, disallow OPTIONS request on pipelines all pipelines except

1) The pipelines handled by CXF
2) The pipelines delegating to another web application
3) The PageModelApi pipeline (which really should allow the OPTIONS
request to indicate what is allowed according the server)

Nest to that, setting additional defaultResponseHeaders for the
CorsSupportValve was not be needed (and complicating the logic): instead
just using a boolean corsAllowCredentials for adding/setting the
Access-Control-Allow-Credentials header is better