HSTTWO-4717 Load platform Module as the first Spring Addon Module
Description
Typically, SpringComponentManager-platform.xml contains Spring beans
which register some service

These services typically are the implementation of a shared api
interface, and as such, can be used cross webapp.

These services can also be used by downstream projects.

Now, it becomes a bit nasty if such a service is used by Spring Bean
initialization in downstream projects in a Spring Addon Module: The
problem is that SpringComponentManager-platform.xml itself is a spring
addon module, and might be loaded later than the Spring addon module
that wants to access a service registered by
SpringComponentManager-platform.xml : To avoid the service not yet
being present, the access of the service has to be done lazily.

It would be much handier if the SpringComponentManager-platform.xml
would always be the first spring addon module to load