HSTTWO-4712 Passwords cannot be reused for new logins
After a login a password gets nullified to avoid it being available in
memory dumps.

Therefor, for a session based mount which renders the page with the
session of the currently logged in user, you cannot use the supplied
credentials to login again (only with the exact same credentials but
not with a new credentials object derived from the used one)

This makes it really hard to use disposable sessions for session based
mounts. To overcome this, we can use impersonate for LazySessions, and
make sure that any session created in a LazySession via impersonation
can be logged out explicitly