HSTTWO-4719 Explicit allow origin for request methods other than OPTIONS
Until now, only for the OPTIONS request we'd potentially change the
response header

Access-Control-Allow-Origin: *

into an explicit origin, for example

Access-Control-Allow-Origin: https://www.example.org

This is required for XMLHttpRequest with attribute withCredentials. It
turns out that this however is also needed for requests other than
the OPTIONS request. Therefor, also for for example a GET Page Model API
request we have to modify the * into an explicit origin