CMS-11846 Convert system plugins to perspectives
By converting plugins into perspectives we ensure that the
CMS calls updateNavLocation on the parent so that the
navapp updates browser url and breadcrumb. As long as there
are no cross-links between perspectives or deep-links to
elements in a list this is fine. However, the users, groups,
user roles and permissions plugins have cross reference links
between them. As a consequence it is not so easy to convert
these plugins into perspectives. This would require javascript
code to switch perspectives and deep linking (e.g. to a role).

The system info, system properties and system updater plugins
don't have cross links to other plugins. After consulting with
PM, for now we decided to only convert these system plugins
into panels and keep the user management plugins on the admin
perspective.