HSTTWO-4716 Use Accept-version as request header and 400 for unsupported
Instead of using API-Version as response and request header, use
Accept-version as request header. Also make sure that for OPTIONS request,
'Accept-version' is an accepted header according
Access-Control-Allow-Headers.

If a request is done for an API version that is not supported / present,
return a 400 BAD_REQUEST instead of default page model api response