HSTTWO-4716 Support an end project to define extra CORS allowed headers
Since an end project / SPA would like to send headers the HST is not
aware of, support a hook to inject extra allowed headers via
platform.properties.

The property to use is

extra.access.control.allowed.headers

and the value must be a comma separated list of supported extra
headers. By default, we use

${extra.access.control.allowed.headers:} implying there are no extra
headers (The ':' makes sure the default value when missing is an
empty String)