CMS-11846 Get rid of unused breadcrumb panel
The three system panels do not interact with the breadcrumb.
Therefore I introduced a simple SystemPanel that the other
system panels can extend. I also tweaked the layout a bit to
make it look nicer.

Note that the UpdaterPanel has a feedback component because it
used to inherit it from PanelPluginBreadCrumbPanel. Seems like
it's never used however, but I'm keeping it in just to be sure.