HSTTWO-4726 Support stateless requests validation
When a request is for the Page Model API, it should never create an
http session: typically the PMA request is done from XHR requests
without cookies, hence creating an http session is pointless and adds
many newly created http session which only get invalidated after,
say 30 minutes when the http session expires.

This is very undesirable.

Therefor, start logging an ERROR in case an new http session is
created while the request is for a Page Model API request. Also log
the stack that created the http session

And finally, make sure to invalidate the session again since the PMA
requests should not leave http sessions behind

Since this improvement is also meant for 14.2.0, we by default keep the
old behavior by not logging an ERROR if http sessions are created since
this might be an annoyance for existing projects that start to log
errors (for example if some forge project component creates (faulty)
an http session). By setting

hst.stateless.request.validation = true

the HST will log errors and log out a newly create http session if the
request was for a stateless resource (like Page Model API)

The issue is related to HSTTWO_4597, since that issue introduced
supporting to serve the PMA preview for the channel manager over the
host of the live website without http session but only a token