HSTTWO-4597 Support custom serverid and add CORS headers
The 'serverid' is configurable now via property
cluster.node.affinity.param and the default is:

cluster.node.affinity.param = serverid

It is the param that is used for cluster node affinity,
and by default we use serverid for it. For this reason
also changed #getServerId into the more general method
#getClusterNodeAffinityId. While at it, moved #getOrigin
to HstRequestUtils from CorsSupportValve

Also make sure that serverid is an allowed custom
CORS header, just like jwt.token.authorization.header now
is also an allowed custom header. Note that the values
are taking from the platform webapp properties, so
if they are (ever!) changed by an endproject, they have
to be changed in the hst and platform properties to the
same value. Not super clean but I consider acceptable since
unlikely to be needed any time soon

Finally the redirect logic in the HstFilterDelegateeBean
to a preview URL has been hardened. By moving the
logic below the decoration of the mount to a preview mount,
it now also takes the preview URL config from the preview
channel node instead of from the live channel node