HSTTWO-4597 'Authorization' needs to be present as CORS allowed header
From [1] we can read:

"The value "*" only counts as a special wildcard value for requests
without credentials (requests without HTTP cookies or HTTP authentication
information). In requests with credentials, it is treated as the
literal header name "*" without special semantics. Note that the
Authorization header can't be wildcarded and always needs to be
listed explicitly."

So even if any header is allowed, Authorization still needs to be added
explicitly

[1] https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Headers