HSTTWO-4734 Replace Jackson Mixin logic with Page Model specific annotations
Instead of relying heavily on the mixin logic, which makes serialization
more complex and also more brittle due to potential new methods in
HippoBean which break serialization (if not accounted for in all of the
serialization processes using Mixins), we now have dedicated annotations
for HippoBean serialization. Since the serialization might be different
for Page Model API and for example Content feed, the annotations
introduced are specifically for serializing the HipoBeans into the
Page Model API output. Note that some extra fields are now serialized,
for example for images, like the size of the image. Adding fields is
BC