HSTTWO-4736 Include generic 'learning' skyscreamer JSONAssert tests
From the ApiCompatibilityJsonAssertTest it became clear that skyscreamer
JSONAssert is really nice for validating that the Page Model API does
not change backwards incompatible. There are different comparison modes
with skyscreamer, where it turns out that for our Page Model API we
should use JSONCompareMode.STRICT_ORDER : this mode namely assures that
extra fields or objects are allowed to be added, and fields and objects
are allowed to be reordered. However field values are not allowed to
change, neither array ordering, and lastly, fields are not allowed to be
removed