CMS-12522 Add folder sorting mechanism
If the mixin is set on a hippostd:folder node then
the FolderTreeNode will return it's child folder
nodes sorted in alphabetical order.

Override sortChildTreeNodes implementation

Couldn't find out why, but the childComparator from the
super class JcrTreeNode is not used for sorting the
child folder nodes. So instead I implement sorting by
overriding FolderTreeNode#sortChildTreeNodes. To prevent
sorting in the console I had to move the constants "cms"
and "console" to the api module (not so nice, but hey,
there's more in api that doesn't belong there).

Add and use folder sorting mechanism icon.
Use grid css styling on the parent div, because the wicket generated
html is hard to restyle.
Introduce adapter to use with wicket PropertyModel to prevent creating a
POJO based on the folder and then setting the mixin again.
Rename translation keys to reflect the latest functional terminology:
Set folder sorting mechanism.