CHANNELMGR-2454 Restyle channel&page fullscreen dialog header
Moved save&cancel form buttons into the header for page-new, page-copy,
page-move, page-properties and channel-settings subpages. This means
the close button is gone. Functionally, the close button triggers the same
behavior as the 'cancel' button, but it will require some work in the
E2E tests as the buttton with class "qa-button-back" is gone.