HSTTWO-4724 Implementation of 'flattened' Page Model API
Now based on 'known' PMA entities, the Page Model output gets
flattened. Thus instead of deeply nested component structure for example,
the structure is flattened and referenced via $ref.

Getting the Jackson serialization right, certainly the PageModelSerializer
was very tough...

Note that not everything is covered yet. The content blocks example does
not get serialized correctly, and we need to still account for max
depth of child content beans to be included in the result