CMS-12527 Skip CM iframe navigation if iframe is not registered
Channel Manager's iframe is registered with some delay. In a case of slow network connection user is able to click CM menu button before the iframe is registered so an attempt to navigate there leads to errors. Existing check with isFirstNavigation only works for one click but if there are series of subsequent clicks that doesn't help.

The solution simply ignores internal iframe navigation if iframe is not defined.