CHANNELMGR-2480 Fix error when CM is opened first time
The source of the error in the active digest cycle in Firefox. In the same time digest cycle isn't active the exactly same moment. Attempts to find the real reason of that behavior were unsuccessful. The simple fix isn't ideal but it fixed the problem without long investigation. Since the framework (angular.js) is outdated that should be enough for now.