HSTTWO-4736 Update expectation files AND validate ImageSetBean serialization
Validate imageset bean serialization with type = 'imageset'

Note that the imageset serialized now contains

"links":{
        "site":{
          "href":"/site/spa/pagenotfound",
          "type":"internal"
        }
},

and then the resources like original and thumbnail that contain a
working href! The 'links' for imageset bean itself AND for assetbean
does not mean anything since cannot be resolved to a sitemap item at
all (in general!). Hence instead, we should not return a links object
for imageset and asset. This will be picked up in issue HSTTWO-4739