HSTTWO-4725 Correct the 'SELF' links for v0.9 and v1.0
in v0.9 only 'external' field will be added and the wildcards will be
resolved (was really a bug). For the rest no changes since the changes
are incompatible (so not even including querystring)

In v1.0, we do return 'external' field, resolve the wildcards, include
the query string and also make sure that in case of partial page
rendering, the SELF link contains the partial page info (but not the
SITE link because SPA does not 'know' partial URL logic). For example
a URL like http://localhost/site/api./header?dummy=bar will result in
something like this in v1.0

{
  "meta" : {
    "version" : "1.0"
  },
  "links" : {
    "site" : {
      "href" : "/?dummy=bar",
      "type" : "internal"
    },
    "self" : {
      "href" : "http://localhost/site/api./header?dummy=bar",
      "type" : "external"
    }
  },
}

Note as part of this commit, to the SITE link also the query string is
now added in v1.0